stage_2/requirements.txt
kafka-python
elasticsearch==8.12.0
pymongo